Implemented scraper result cache from db.

+++ src/main/java/com/vhausler/property/stats/model/dto/ParameterDTO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class ParameterDTO {
-    private int id;
-    private int scraperResultId;
+    private long id;
+    private long scraperResultId;
     private String key;
     private String value;
 }
